deleted AddMealModal and renamed Schedule file #71
